Start 'kubectl proxy' with random port and Bash process substitution
The consequence of this is that multiple instances of the dashboard can
be run independently from each other (before, this was not possible
because the instances would try to use the same harcoded port).

To parse the chosen random port, the 'kubectl proxy' process is started
with Bash process subsitution '<()' rather than the ampersand '&'
(solution in [*]).

This has in turn another consequence: when the dashboard is exited with
Ctrl-C, the 'kubectl proxy' process is deleted too. This means, that the
capturing of the SIGINT signal and the explicit killing of the 'kubectl
proxy' process is not necessary anymore.

In other words, now, when exiting the dashboard with Ctrl-C, *all* the
associated processes are deleted. Before, the 'kubectl proxy &' process
(started with '&') was not deleted and thus needed to be explicitly
deleted in the 'cleanup' function.

[*] https://stackoverflow.com/a/20018118/4747193