Check for readiness of Pods
When a Pod is terminating, its Ready condition is false, but its phase
is still Running. Now, the code checks for the status of the Ready
condition, and if it's false, but the phase is Running, the phase is
changed to Pending.

This improves the UX when terminating Pods (e.g. scaling down a
Deployment). The Pods that get terminated are now immediately marked as
Pending until they are fully deleted (before, they just stayed Running
they were fully deleted).