Add support for specifying resource type
The resource type (e.g. pods, deployments) can be specified as the
second command-line argument (after the namespace).

At the beginning, the code makes a 'kubectl get' request for this resource
type, which serves two purposes:

  1. Check if it's a valid resource type. If the resource type is invalid
     (e.g. "foo"), the 'kubectl get' command will fail and k1s will exit
     with an error message too.

  2. Identify the API group and version of the specified resource type.
     This information is extracted from the output of kubectl (with
     increased verbosity). This is informaiton is needed to construct
     the URL for the watch request. kubectl includes all the logic for
     identifying this information (it also caches it), so it makes sense
     to use kubectl. In fact, the entire API path is extracted from the
     kubectl output and directly used in the watch request.

A consequence of the last point above is that now also aliases of
resource type names can be specified. For example, for Pods, "pods",
"pod", and "po" can be specified, and for Deployments, "deployments",
"deployment", and "deploy" can be specified, and so on.