Add current and required replica count to Pod controllers
Deployments, ReplicaSets, and StatefulSets now display the currently
ready number of replicas and the required number of replicas (for
example, "5/5").

This is interesting for scaling and updating Deployments. For example,
it allows to perform a rolling update on a Deployment and observe how
the actual number of replicas stays withing certain range (defined by
maxSurge and maxUnavailable).