Add support for "all namespaces" and non-namespaced resources
If the namespace command-line argument is "-" and a namespaced resource
type is specified, then the resources across all namespaces are
displayed.

If the namespace argument is "-" and a non-namespaced resource type is
specified (e.g. clusterroles), then all the resources of this type are
displayed.

If a non-namespaced resource type is specified (e.g. clusterroles) and
the namespace command-line argument is not "-" (e.g. it's "default"),
then this is detected and the namespace indicator in the dashboard
banner is set to "-" (this detection works by counting the number of
elements in the detected API path; if it's 3 or 4, the path does not
include a namespace, if it's 5 or 6, the path includes a namespace).