Revert the order of dependency check and version output
It's ok if the version can be printed even if the dependencies are not
satisfied. Also allows `brew test k1s` to pass even if not all
dependencies are installed.

Note about the Homebrew formula being created:

Probably not declaring the dependencies on jq, watch, curl, and kubectl
in the Homebrew formula, since these would create dependencies on the
Homebrew formulas of these tools. However, the k1s formula should also
install if these dependencies are installed in another way. Another
disadvantage of declaring dependencies on other Homebrew formulas is
that these referred formulas then can't be uninstalled without also
unistalling the formula that depends on them.