Support both GNU sed and BSD sed (fixes #2)
Two of the sed commands previously worked only for GNU sed:

  1) sed 's/.*:\([0-9]\{4,5\}\)\b.*/\1/'
  2) sed -i <...>

The problem with (1) was the \b (word boundary) escape sequence, which
is not understood by BSD sed. It's not really necessary, and removing it
solves the problem.

The problem with (2) was that with BSD sed, the -i option (in-place file
editing) requires an extension for the backup file to create. GNU sed
also suppports extensions for the -i option, so adding an extension (e.g.
'-i.bkp') solves the problem. However, note that there must be no space
between -i and the extension because while BSD sed accepts both '-i .bkp'
and '-i.bkp', GNU sed only accepts '-i.bkp' (for an empty extension,
i.e. no backup file, BSD sed accepts only [-i ''] and GNU sed accepts
only [-i''], so there is no overlap there, and thus it's not possible to
specify an empty extension in a portable way).

BSD sed is installed by default on macOS and GNU sed is installed by
default on Linux.