Add "release" GitHub Actions workflow
The workflow is triggered when a new release is created. The workflow
uploads the distribution archive with the corresponding checksum to the
release on GitHub.

To do so, the workflow downloads the source code archive of the release
(which is added automatically by GitHub), removes the uneeded file from
it (images, etc.), and uploads the resulting archive as an asset to the
release (the checksum is also uploaded as a separate file).

The distribution archive that this workflow uploads is used by users who
manually install k1s, as well as by the k1s Homebrew formula.

The procedure for creating a new release of k1s is now:

  1. Create a new release on GitHub (specify tag, release title, and
     release description)
  2. Update the k1s Homebrew formula in weibeld/homebrew-core to use the
     new release (update distribution archive URL and checksum)

Between these steps, the distribution archive and checksum is
automatically created by the worfklow.

There are two additional opportunities for automation in the same
workflow: 1) check whether the output of `k1s -v` matches the tag of the
release (it's easy to forget updating this when creating a new release),
and 2) update the k1s Homebrew formula in weibeld/homebrew-core after
creating the distribution archive for the new release. These jobs would
need to run sequentially.