Extend "release" workflow
The workflow now also:

- Checks if the output of `k1s -v` matches the tag name of the release
- Triggers a version bump of the k1s formula in weibeld/homebrew-core

All the tasks of this workflow run in the same job because they share
certain data. This removes the need to use artifacts or the cache to
share this data, which might be overkill.

The version bump of the k1s formula is achievd by triggering a separate
workflow in the weibeld/homebrew-core repository [1] with a "repository
dispatch" API call and event [2].

The repository dispatch API call cannot be made with the GITHUB_TOKEN
secret that's automatically provided in every workflow (this token
allows only access to the repository in which the workflow runs). For
this reason, a "personal access token" [3] must be used for this API
call. The personal access token is added as a secret [4] to this
repository and referenced from the workflow by its name. The name of the
secret is PERSONAL_ACCESS_TOKEN.

[1] https://github.com/weibeld/homebrew-core/commit/2502e61442cda64505ac88d15fa6ccbb7db72ff5
[2] https://docs.github.com/en/rest/reference/repos#create-a-repository-dispatch-event
[3] https://docs.github.com/en/github/authenticating-to-github/creating-a-personal-access-token
[4] https://docs.github.com/en/actions/configuring-and-managing-workflows/creating-and-storing-encrypted-secrets